+++ crates/svg2pdf-cli/src/args.rs
